добавил структуру компонентов страниц, перенес по семантике
