+++ src/pages/Main/styles.module.scss
@@ -1,0 +1,1 @@
+@use "../../assets/styles/variables";
