+++ src/pages/Topic/Topic.tsx
@@ -1,0 +1,7 @@
+import styles from './styles.module.scss';
+
+const Topic = () => {
+  return <div></div>;
+};
+
+export default Topic;
